virelo-ai-worker/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.3
python-multipart==0.0.6
requests==2.31.0
psd-tools==1.9.30
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
pytoshop
